refactor(options): refactor options resolving to fix server options
The server Sentry config was not always consistent. The resolved config
might have been a bit different depending on if using nuxt dev or
nuxt start.
