+++ lib/module.js
@@ -105,8 +105,18 @@ export default function SentryModule (moduleOptions) {
 
   this.nuxt.hook('build:before', () => buildHook(this, options, logger))
 
-  if (serverSentryEnabled(options)) {
-    this.nuxt.hook('ready', () => initializeServerSentry(this, options))
+  // This is messy but Nuxt provides many modes that it can be started with like:
+  // - nuxt dev
+  // - nuxt build
+  // - nuxt start
+  // - nuxt generate
+  // but it doesn't really provide great way to differentiate those or enough hooks to
+  // pick from. This should ensure that server Sentry will only be initialized **after**
+  // the build was done (so that the release version is determined).
+  // It won't be initialized when generating.
+  const initHook = this.options._start ? 'ready' : 'build:done'
+  if (initHook && serverSentryEnabled(options)) {
+    this.nuxt.hook(initHook, () => initializeServerSentry(this, options, logger))
     this.nuxt.hook('generate:done', () => shutdownServerSentry())
   }
