+++ lib/plugin.client.js
@@ -3,10 +3,10 @@ import merge from 'lodash.mergewith'
 import * as Sentry from '@sentry/browser'
 <%
 if (options.initialize) {
-  let integrations = options.SENTRY_PLUGGABLE_INTEGRATIONS.filter(key => key in options.integrations)
+  let integrations = options.PLUGGABLE_INTEGRATIONS.filter(key => key in options.integrations)
   if (integrations.length) {%>import { <%= integrations.join(', ') %> } from '@sentry/integrations'
 <%}
-  integrations = options.SENTRY_BROWSER_INTEGRATIONS.filter(key => key in options.integrations)
+  integrations = options.BROWSER_INTEGRATIONS.filter(key => key in options.integrations)
   if (integrations.length) {%>const { <%= integrations.join(', ') %> } = Sentry.Integrations
 <%}
 }
