+++ lib/plugin.lazy.js
@@ -114,10 +114,10 @@ async function loadSentry (ctx, inject) {
   const Sentry = await import(/* <%= magicComments.join(', ') %> */ '@sentry/browser')
   <%
   if (options.initialize) {
-    let integrations = options.SENTRY_PLUGGABLE_INTEGRATIONS.filter(key => key in options.integrations)
+    let integrations = options.PLUGGABLE_INTEGRATIONS.filter(key => key in options.integrations)
     if (integrations.length) {%>const { <%= integrations.join(', ') %> } = await import(/* <%= magicComments.join(', ') %> */ '@sentry/integrations')
 <%  }
-    integrations = options.SENTRY_BROWSER_INTEGRATIONS.filter(key => key in options.integrations)
+    integrations = options.BROWSER_INTEGRATIONS.filter(key => key in options.integrations)
     if (integrations.length) {%>  const { <%= integrations.join(', ') %> } = Sentry.Integrations
 <%}
   %>
