+++ package.json
@@ -23,7 +23,9 @@
   "types": "types/index.d.ts",
   "scripts": {
     "dev:fixture": "nuxt -c ./test/fixture/default/nuxt.config.js",
-    "dev:generate": "nuxt generate -c ./test/fixture/default/nuxt.config.js",
+    "dev:fixture:build": "node ./node_modules/nuxt/bin/nuxt.js build -c ./test/fixture/default/nuxt.config.js",
+    "dev:fixture:start": "node ./node_modules/nuxt/bin/nuxt.js start -c ./test/fixture/default/nuxt.config.js",
+    "dev:generate": "nuxt generate -c ./test/fixture/default/nuxt.config.js --force-build",
     "lint": "eslint --ext .vue,.js,.ts lib test types",
     "lint:fixture": "eslint --ext .vue,.js --no-ignore 'test/fixture/*/.nuxt/sentry.*'",
     "release": "release-it",
